% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plsgmd.R
\name{plsgmd}
\alias{plsgmd}
\title{Partial least squares with the generalized matrix decomposition}
\usage{
plsgmd(X, Y, H = diag(nrow(X)), Q1 = diag(ncol(X)), Q2 = diag(ncol(Y)), k = 2)
}
\arguments{
\item{X}{The n by p data matrix of predictors}

\item{Y}{The n by q matrix of outcomes}

\item{H}{An n by n sample-wise similarity matrix.}

\item{Q1}{A p by p similarity matrix of the features in X.}

\item{Q2}{A q by q similarity matrix corresponding to columns in Y.}

\item{k}{Integer, the number of PLS latent variables to compute.}
}
\value{
Scores and loadings for X and Y.
}
\description{
Partial least squares with the generalized matrix decomposition
}
